Fix issues in the code
