+++ classroom.rb
@@ -11,5 +11,6 @@ class Classroom
 
   def add_student(student)
     @students << student
+    student.classroom = self
   end
 end
